THE-554 split manager chunk io files
